+++ src/main/java/guru/qa/service/impl/PowOperationHandler.java
@@ -5,8 +5,8 @@ import guru.qa.service.OperationHandler;
 public class PowOperationHandler implements OperationHandler {
 
     @Override
-    public double invoke(int first, int second) {
-        return Math.pow(first, second);
+    public int invoke(int first, int second) {
+        return (int) Math.pow(first, second);
     }
 
 }
